fix: update apps_view to use app_media
